Factor out BlobStats from collector, add a new AddBlobs method
